First Iteration of Writer Object
Working on a terminal to display text. Currently only displays some
sample text, but will eventually display information based on which room
you are in. Test room is being used for the demo. The step function will
also have to divide the string into an array of words to be conscious of
the right hand side of the terminal to make indents, but I'm tired and
half drunk so I'll work on that later.
There is also a test of character writing sounds. That is still in the
works, as a constantly repeating sound can get really annoying real
quick.
Multi-page functionality will be in the works once I finish the
indentation process.
Anyways, this is a good place to start. These text scrollers can get a
complicated quick so I just wanted to start with the basics and see what
to tackle next.