Second Iteration of Writer Object: Broken Multiline Boogoloo
Started the process of the writer class, on creation, breaking a string
down into multiple lines stored in an array. The animation part isn't an
issue, as that can be added later on. The str_copy method is messed up
and is carrying duplicate string information per line. It probably has
something to do with the lineIndex variable in the Create part of the
writer object. I just need time to think about this and come right back
to it. Can't wait to start doing multiple pages. That will be fun.