ScCheckRoomPages
Basically a setup for adding multipage functionality. A global variable
pageList is reinstantiated every time a writer object is found in a
room. Then, it is filled with strings that are specific to the room. I
was hoping the creation code could hold variables, but this will have to
do. next sync I'll try to have the Writer object itself allow the
multiple strings to be displayed per page.